Se implementaron los casos de prueba de la clase Sanguchetto, SanguchettoTest.
+ public void testVaciar()
+ public void testAgregarIngrediente()
+ public void testVerIngredientes()
+ public void testVerCondimentos()
+ public void testGetPrecio()

+++ src/main/java/tallerweb/sangucheto/modelo/Sanguchetto.java
@@ -62,7 +62,7 @@ public class Sanguchetto {
 	 * @return
 	 */
 	public Double getPrecio(){
-		Double precioTotal = (double) 0;
+		Double precioTotal = 0.0;
 		
 		for (Ingrediente cadaElemento : ingredientes)
 			precioTotal += cadaElemento.getPrecio();
